feat: user follower refactor + email notifications for inactive users
Made user followers a separate class from the user class (thank god) so I had to adjust a lot of the following logic.
Added lastLogin to the user class so I had to modify the register method too.
Added a sendWeeklySummaries method that send an email to inactive  users. Inactive is anyone who's last login was 7 days ago